+++ sdk/storage/Azure.Storage.DataMovement/tests/Shared/StorageResourceTestWrappers.cs
@@ -69,8 +69,8 @@ public class StorageResourceItemFailureWrapper : StorageResourceItem
     protected internal override Task<HttpAuthorization> GetCopyAuthorizationHeaderAsync(CancellationToken cancellationToken = default)
         => ThrowOrDo(() => _inner.GetCopyAuthorizationHeaderAsync(cancellationToken));
 
-    protected internal override StorageResourceCheckpointData GetDestinationCheckpointData()
-        => ThrowOrDo(_inner.GetDestinationCheckpointData);
+    protected internal override StorageResourceCheckpointDetails GetDestinationCheckpointDetails()
+        => ThrowOrDo(_inner.GetDestinationCheckpointDetails);
 
     protected internal override Task<string> GetPermissionsAsync(StorageResourceItemProperties properties = null, CancellationToken cancellationToken = default)
         => ThrowOrDo(() => _inner.GetPermissionsAsync(properties, cancellationToken));
@@ -78,8 +78,8 @@ public class StorageResourceItemFailureWrapper : StorageResourceItem
     protected internal override Task<StorageResourceItemProperties> GetPropertiesAsync(CancellationToken token = default)
         => ThrowOrDo(() => _inner.GetPropertiesAsync(token));
 
-    protected internal override StorageResourceCheckpointData GetSourceCheckpointData()
-        => ThrowOrDo(_inner.GetSourceCheckpointData);
+    protected internal override StorageResourceCheckpointDetails GetSourceCheckpointDetails()
+        => ThrowOrDo(_inner.GetSourceCheckpointDetails);
 
     protected internal override Task<StorageResourceReadStreamResult> ReadStreamAsync(long position = 0, long? length = null, CancellationToken cancellationToken = default)
         => ThrowOrDo(() => _inner.ReadStreamAsync(position, length, cancellationToken));
@@ -155,10 +155,10 @@ public class StorageResourceContainerFailureWrapper : StorageResourceContainer
     protected internal override Task CreateIfNotExistsAsync(CancellationToken cancellationToken = default)
         => ThrowOrDo(() => _inner.CreateIfNotExistsAsync(cancellationToken));
 
-    protected internal override StorageResourceCheckpointData GetDestinationCheckpointData()
-        => ThrowOrDo(_inner.GetDestinationCheckpointData);
+    protected internal override StorageResourceCheckpointDetails GetDestinationCheckpointDetails()
+        => ThrowOrDo(_inner.GetDestinationCheckpointDetails);
 
-    protected internal override StorageResourceCheckpointData GetSourceCheckpointData()
-        => ThrowOrDo(_inner.GetSourceCheckpointData);
+    protected internal override StorageResourceCheckpointDetails GetSourceCheckpointDetails()
+        => ThrowOrDo(_inner.GetSourceCheckpointDetails);
     #endregion
 }